requirements.txt
flask
flask-smorest
python-dotenv
Flask-SQLAlchemy
flask-jwt-extended
passlib
flask-migrate